requirements.txt
python-telegram-bot~=21.10
psutil~=5.9.8